move RegisterCMap to own file
